Implement P1 core audio boundary refactor
